chore: moved the logging stuff to it own file
