+++ engine/src/main.rs
@@ -91,7 +91,7 @@ async fn main() -> anyhow::Result<()> {
 
 	task_scope(|scope| {
 		async move {
-			let mut start_logger_server_fn = Some(utilities::init_json_logger(settings.logging.clone()).await);
+			let mut start_logger_server_fn = Some(utilities::logging::init_json_logger(settings.logging.clone()).await);
 
 			let ws_rpc_client = jsonrpsee::ws_client::WsClientBuilder::default()
 				.build(&settings.state_chain.ws_endpoint)
